Move to Kotest 6.0.M1 + new `arrow-platform`
