+++ arrow-libs/core/arrow-atomic/src/commonTest/kotlin/arrow/atomic/AtomicTest.kt
@@ -1,6 +1,7 @@
 package arrow.atomic
 
 import arrow.fx.coroutines.parMap
+import arrow.platform.stackSafeIteration
 import kotlin.test.Test
 import kotlinx.coroutines.test.runTest
 import io.kotest.matchers.shouldBe
