+++ arrow-libs/core/arrow-core/build.gradle.kts
@@ -37,6 +37,7 @@ kotlin {
     commonTest {
       dependencies {
         implementation(projects.arrowFxCoroutines)
+        implementation(projects.arrowPlatform)
         implementation(libs.kotlin.test)
         implementation(libs.coroutines.test)
         implementation(libs.kotest.assertionsCore)
