+++ arrow-libs/core/arrow-functions/build.gradle.kts
@@ -33,6 +33,7 @@ kotlin {
     commonTest {
       dependencies {
         implementation(projects.arrowFxCoroutines)
+        implementation(projects.arrowPlatform)
         implementation(libs.kotlin.test)
         implementation(libs.coroutines.test)
         implementation(libs.kotest.assertionsCore)
