+++ arrow-libs/core/arrow-platform/build.gradle.kts
@@ -8,7 +8,6 @@ plugins {
   id(libs.plugins.kotlin.multiplatform.get().pluginId)
   alias(libs.plugins.arrowGradleConfig.kotlin)
   alias(libs.plugins.publish)
-  alias(libs.plugins.kotlinx.kover)
   alias(libs.plugins.spotless)
 }
 
@@ -18,21 +17,13 @@ spotless {
   }
 }
 
+apply(from = property("ANIMALSNIFFER_MPP"))
+
 kotlin {
   sourceSets {
     commonMain {
       dependencies {
-        api(projects.arrowCore)
         implementation(libs.kotlin.stdlib)
-        implementation(libs.coroutines.core)
-      }
-    }
-    commonTest {
-      dependencies {
-        implementation(projects.arrowFxCoroutines)
-        implementation(projects.arrowPlatform)
-        implementation(libs.kotlin.test)
-        implementation(libs.coroutines.test)
       }
     }
   }
@@ -40,7 +31,7 @@ kotlin {
   jvm {
     tasks.jvmJar {
       manifest {
-        attributes["Automatic-Module-Name"] = "arrow.resilience"
+        attributes["Automatic-Module-Name"] = "arrow.platform"
       }
     }
   }
@@ -50,8 +41,4 @@ kotlin {
     (project.rootProject.properties["kotlin_language_version"] as? String)?.also { languageVersion = KotlinVersion.fromVersion(it) }
     (project.rootProject.properties["kotlin_api_version"] as? String)?.also { apiVersion = KotlinVersion.fromVersion(it) }
   }
-}
-
-tasks.withType<Test>().configureEach {
-  useJUnitPlatform()
 }
