+++ arrow-libs/core/arrow-platform/gradle.properties
@@ -1,0 +1,2 @@
+# Maven publishing configuration
+POM_NAME=Arrow Platform
