+++ arrow-libs/core/arrow-platform/knit.properties
@@ -1,0 +1,5 @@
+knit.package=arrow.platform.examples
+knit.dir=src/commonTest/kotlin/examples/
+
+test.package=arrow.platform.examples.test
+test.dir=src/commonTest/kotlin/examples/autogenerated/
