+++ arrow-libs/core/arrow-platform/src/jsMain/kotlin/arrow/platform/Platform.kt
@@ -1,0 +1,3 @@
+package arrow.platform
+
+public actual val platform: Platform = Platform.JS
