+++ arrow-libs/core/arrow-platform/src/jvmMain/kotlin/arrow/platform/PlatformJvm.kt
@@ -1,0 +1,4 @@
+@file:JvmName("CurrentPlatform")
+package arrow.platform
+
+public actual val platform: Platform = Platform.JVM
