+++ arrow-libs/fx/arrow-fx-coroutines/build.gradle.kts
@@ -34,6 +34,7 @@ kotlin {
       dependencies {
         implementation(projects.arrowCore)
         implementation(projects.arrowAtomic)
+        implementation(projects.arrowPlatform)
         implementation(libs.kotlin.test)
         implementation(libs.coroutines.test)
         implementation(libs.kotest.assertionsCore)
