--- arrow-libs/resilience/arrow-resilience/src/commonTest/kotlin/arrow/resilience/Platform.kt
@@ -1,9 +1,0 @@
-package arrow.resilience
-
-import arrow.resilience.common.Platform
-import arrow.resilience.common.platform
-
-fun stackSafeIteration(): Int = when (platform) {
-  Platform.JVM -> 20_000
-  else -> 1_000
-}
