+++ gradle/libs.versions.toml
@@ -4,7 +4,7 @@ arrowGradleConfig = "0.12.0-rc.24"
 coroutines = "1.9.0"
 classgraph = "4.8.177"
 dokka = "1.9.20"
-kotest = "5.9.1"
+kotest = "6.0.0.M1"
 kover = "0.8.3"
 kotlin = "2.0.21"
 kotlinBinaryCompatibilityValidator = "0.16.3"
